pyproject.toml
[tool.poetry]
name = "backend-apis"
version = "0.1.0"
description = ""
authors = ["admin1282 <gaikwadvinayak1282@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-split-settings = "^1.3.1"
python-decouple = "^3.8"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
pytest-django = "^4.8.0"
celery = "^5.4.0"
redis = "^5.2.0"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
